✨ matchRate 계산 로직 및 TestResult Submit 로직 분리

+++ src/main/java/com/nexters/keyme/domain/test/presentation/dto/response/TestResultResponse.java
@@ -20,7 +20,7 @@ public class TestResultResponse {
     private Long testId;
 
     @ApiModelProperty(value = "테스트 제출자와의 일치율(MMVP에서는 없음)", example = "87.7")
-    private float matchRate;
+    private Double matchRate;
 
     private List<QuestionSolvedResponse> results;
 }
